refactor(classification): 统一分类规则候选选择 (#341)
统一分类规则候选资格、编译与 priority/rule_id 选择，并迁移正式账单刷新与导入 Stage 2。

+++ src/backend/http/import_routes/ocr_learning_handlers.rs
@@ -225,9 +225,9 @@ async fn load_receipt_draft_context(
             .map(|rule| ReceiptDraftCategoryRule {
                 id: rule.id.to_string(),
                 category_id: rule.category_id.to_string(),
-                category_type: rule.category_type,
+                category_type: i64::from(rule.category_type),
                 label: category_label(&rule.main_category, &rule.sub_category),
-                priority: rule.priority,
+                priority: i64::from(rule.priority),
                 rule_expression: rule.rule_expression,
                 regex_enabled: rule.regex_enabled,
             })
